Changed to JSON reservation file.
I didn't like the old way of doing reservations, where I tried to glean info about a reservation from the files created about it.
Instead, I create a single file containing JSON structs which give information about a reservation name, who owns it, when it expires, etc. Significantly simpler and reduces the number of I/O operations.
Downside: It will be possible for the file to become out of sync with what's actually on disk. Unlikely, but possible.