minicli: fixed bug in Prefix.
The command worked when there was a pattern that was a prefix of all the
patterns, for example, "vm info" is the prefix of:
  vm info
  vm info search
  vm info mask
However, if there wasn't such a pattern, such is the case with the vm
inject command:
  vm inject dst ...
  vm inject src ...
Then the returned prefix was empty. Fixed the bug and updated the tests.