minimega: implemented host command.
Implemented the host command which takes the place of the old host_stats
and hostname commands. With a single argument, the host command return
the particular stat. With no arguments, host returns tabular data for
all the stats it knows how to compute. The implementation could be
improved since currently it requires running hostStatsMemory twice to
get total and used memory even though both are returned when running it.