minimega/minicli: implemented the read command.
Implemented the read command which required a completely different
interface for returning responses and recording history. Removed the
record bool from handler, it will be up to the code calling minicli to
determine whether it's appropriate to record a command. Simplified a lot
of the cli handlers using a wrapper in misc.go.