minicli: bug fix, change pattern printing.
Many commands in minimega are written to take advantage of a hack with
minicli's patterns: a required choice field with one choice has the same
effect as a required string but with the extra effect of setting a flag
in BoolArgs. This can simplify the CLI handler greatly. Changed minicli
so that required choice fields with one argument are printed as if they
were a required string since it doesn't matter to the user.

Also, while changing the printing, discovered a bug in the choice code
where an empty string was getting into Options when a comma was followed
by the terminal character.