minimega: initial work on replacing vmConfigFns.
Getting rid of the ugly function pointer structs in vmConfigFns.
Replacing with three generic getters: getString, getBool, and
getStringSlice. These three getters take a field name and return a
pointer to the vmInfo field with the name corresponding to the
parameter. This allows the cli handler to figure out what field it
should update (and how the update should be performed) based on which
getter returns a non-nil pointer.

Also adding tests to make sure that vm config works the way we intend it
to. Specifically, testing whether it sets/clears correctly for every
field.

Still need a bit more work to test append/net/qemu-override and to
implement vmList.save. Also, need to figure out whether we want to keep
the vm config qemu command or use an `external` command instead.