minimega: dirty hacks.
Work arounds for not all VM state being stored in the VM struct.
Fallback on the Tabular data from minicli.Response to populate VM info
fields that are dynamically generated (ip, ip6, cc_active). This code
should be deleted as soon as the state for these is moved into the VM
struct.

Fixes #175