minimega: yet another bug fix.
Problematic use of pointers to copies of structs. Need to double check
all ranges over vm.Networks to make sure that they don't have a similar
issue.

Example of bug:

```
    vals := []string{"a", "b", "c"}

    for _, v := range vals {
        fmt.Printf("%p -- %v\n", &v, v)
    }
```

Example output:

```
0x1040a120 -- a
0x1040a120 -- b
0x1040a120 -- c
```

The problem was that we were passing &v to a goroutine and v was
changing on the next iteration of the loop.