Proper implementation of graphing functionality. VLANs and routers show up as single nodes. Graph structure is generated like this:   1. Data is pulled from the MM server via the /json URL   2. JSON is parsed, machines that are only on a single VLAN are grouped together, while machines with multiple VLANs are treated as routers   3. Links are created for each router. If a VLAN only contains a router, an empty node is created to represent it.
Layout is handled by a d3 force-directed graph. Display of this layout is handled by a WebGL-backed graphing library (ayasdi/grapher).

The /json URL is polled every 1.5s.
The JSON string is only parsed when the response differs from the last one.
Recalculation of the layout only occurs when there is a change that affects graph structure (change to list of nodes or list of links).

Both VLAN and router nodes can be clicked to be selected.
When a VLAN node is clicked, every machine in that VLAN is listed.
When a router node is clicked, every machine within one hop of that router is listed.
The machines listed can be clicked for further detail.

Panning and zooming in/out on the graph are functional.
Nodes can be dragged as well.