minimega: rework screenshotting
This is intended to fix #252. It maintains compatibility with the
old "vm screenshot" command while allowing you to say this:

	vm screenshot 0 file /dev/null
	vm screenshot 0 file /dev/null 100

Instead of writing out the screenshot file at the deepest level, I
pass back a slice of bytes containing the png data, then write
it out in vm_cli's cliVmScreenshot function. Previously, we were
writing to the file, then reading that back out into the response's
Body field, which made it fail when you give /dev/null as a file.