minimega.init: fix ordering for startup check
When starting minimega, it attempts to read the pid file, then checks
if that was successful. Unfortunately someone put a "sleep" between the
read and the check, meaning the check actually checks the return status of
"sleep". I fixed it and made the sleep longer because minimega sometimes
starts a little slowly.