minimega: refactor namespace launching.
Moved code around for how the namespace VM launching works. Instead of
trying to keep things in sync with the `vm config ...` commands across
all the hosts, those commands will only apply locally. When the user
runs `vm launch ...` with a namespace active, this will queue VMs with
those configs. The user will then have to run `vm launch` to actually
launch those VMs with a namespace active. Should not change the
functionality when namespaces are not active.