minimega: allow scripted multi-file VNC playback
In addition to the existing PointerEvent and KeyEvent in VNC playback,
files can now include a LoadFile directive. This will cause the specified
file to be loaded and played. The format follows the usual VNC playback
format, e.g.:

	0:LoadFile,browse_slashdot.vnc
	10000000000:LoadFile,/home/john/recordings/reboot_windows.vnc

The filenames may be relative to the location of the current file,
or they can be absolute. Timestamps are honored as usual.