minitunnel: remove connect race. Fixes #407.
There was a race in minitunnel -- if you forwarded a port and used it
immediately, before the TID was registered, then some of the data would
get lost causing the tests to hang. Moved spawning goroutine to after
registering the TID to fix. Ran tests for ten minutes with no issues.