minimega: fix what recursively set Source broke
Recursively setting Source broke running commands on remote hosts when
invoked via `namespace <namespace> (cmd)`. We need the recusive setting
of Source so, instead of reverting that, we change the check to see if
Source is SourceMeshage in a bunch of places. The assumption is that the
only time we would want a namespace-aware broadcast is when we're
running on the head node. If someone invoked `mesh send ccc1 namespace
foo host`, she will get the host info for ccc1 and not all the nodes in
the namespace. Can't do better than this without introducing another
field in Command (replacing Source with namespace-specific field).