minimega: huge refactoring of bridge code.
Moved the bridge code into a separate package and heavily refactored it.
We now have a single lock in the bridge code that is locked in every
exported bridge function. All the internal bridge functions assume that
the lock is held and that nothing funny is going on with goroutines.
Spent quite a bit of time documenting and making the code consistent.

This should also help with namespace-aware bridges and taps -- we can
create one `bridge.Bridges` per namespace and have decent isolation. We
still need to decide whether and how to provide transparent bridging
across hosts in the namespace.

This has been superficially tested with minitest. I want to do a lot
more testing before we merge this.

Updates #468, #472.