bridge/minimega: probably fix for #540
Created a local err in `Container.launch` when calling `bridges.Get`. If
this failed, we wouldn't register a MAC with the bridge so when we tried
to unregister it, minimega panicked. Move networking code to a separate
function so it's easier to return an error if there's a problem rather
than break out of the loop. Fixed some potential issues in the bridge
code when creating a bridge. Need to do a larger test.

Updates #540.