Add files via upload
SMTP port changed and added some additional statements so that the connection would actually be closed instead of timing out.

SSH, changed the port to a non-standard port to solve a port binding issue

UDP added so that UDP traffic is generated. Necessary additions to display and transmit UDP added to report.go and protonuke.go.