First pass at web console for containers
Import xterm.js to provide a javascript-based console. The minimega web
interface now provides a "Connect" button for containers, which will
connect to a javascript terminal.

The terminal is implemented as a pseudo-terminal, rather than with
three pipes as used previously. If we regularly plan to use thousands
of containers on the same host, we will need to include code in minimega
startup to increase the number of PTYs available on the system.

Currently, this terminal behaves exactly as the domain socket: if you
open the terminal, then close it and re-open, the previous output will
not be re-displayed. This will be implemented later.