minimega: first cut at improved scheduler
Scheduler uses vm config schedule/coschedule and least loaded first to
assign VMs to nodes. Currently measures load based on CPU commit but it
should be easy to add an nsmod command to change that to Memory load or
Network commit (both of which are already implemented).

Completely untested => probably full of bugs.