igor: scheduling into the future
igor now only accepts scheduling requests with a # of nodes, not a list of specific
nodes to use. It will find the first available slot for them in the schedule.

igor show isn't yet aware of this, so the rack display may be confused until I
overhaul it; however, the schedule listing at the bottom is still accurate.