minimega: possibly the largest refactoring
Changing the way namespaces work so that the active namespace is passed
into the cli* func. Moved as much state as possible to the Namespace
struct so that it changes with the namespace. This simplifies the
different components because they don't need to be namespace-aware --
they simply grab whatever state they need from the Namespace they are
handed. Also made it so that we're in the `minimega` namespace by
default. This is still a WIP but we now pass many tests (after adding
`vm launch` commands).