minimega: 'fix' container scrollback
Container scrollback was buggy; if you ran e.g. vi, then opened
a new xterm.js tab, it would print ";80R" to each open terminal.
Wireshark indicates that xterm.js was sending \033[24;80R when
it connected. I have commented out the line that was sending
that and it seems to work fine without it. This may merit
further examination but for now, it works.