protonuke: add size parameter to image.png
Can now request an image of an arbitrary size from protonuke. protonuke
stores the images and will provide the same image on future requests.
This means that there's a delay on the first request while the image is
generated but future requests will be fast. We keep all images so it is
possible to get OOM by requesting too many different sizes.