meshage: use instance IDs, single random source
Create a new instance ID when the node starts so that we never restart
minimega and have it be unable to connect because the lollipop length
has not been reached.

Needed a random number in NewNode so added rand to the Node struct and
use it for all random numbers rather than reseed everytime.

Fixes #973