ron/minimega: refactoring to redial on Start
Refactor minimega so that we connect CC when starting a VM. Add test to
ensure that we can reconnect after VMs are killed and restarted.

Add additional checks in ron to make sure that we aren't trying to
listen on the same thing twice and properly cleaning up the listeners
that we have closed.

Suppress tunnel EOF messages.