minimega,vnc: prevent crash when injecting WaitForIt
Inject assumed that if it wasn't a pointer or mouse event then it was a
string causing a crash. Removed old functionality so that now it has to
explicitly be a LoadFile event.

Users can now inject a WaitForIt event if there is a running playback
but not if there is not one.