Add -vlan flag to sub and edit
Allows users to specify which vlan they'd like to connect to. The
value of -vlan can be the name of another reservation (the new
reservation uses the VLAN of the existing one) or a VLAN ID (as long
as that VLAN is not in use, or is in use by a reservation that is
writable to the user).