Support miniccc reconnection over serial port
React to async QMP virtual serial port change events, which are
triggered when the miniccc client connects to the virtual serial port in
a VM.

When a connection is opened client-side, the server connects to the
virtual serial port and waits for the initial handshake message.

Client-side, miniccc has been updated to wait after connecting to the
virtual serial port to give the server time to react to the QMP event
and make it's connection to the serial port first.