Use "actual path" from qemu-img for backing image if present
If you run `qemu-img info` in the same directory as the image your
passing to it, then the `(actual path: ...)` stuff isn't included in the
output. However, if you run `qemu-img info` from a different directory,
or you pass the absolute path of the image to `qemu-img info`, then the
`(actual path: ...)` stuff is present. Minimega always passes the full
path of a disk image to `qemu-img info`.