Support TCP and UDP connectivity testing from miniccc
minimega, miniccc, and ron were updated to support executing tcp/udp
connectivity tests directly from miniccc agents.

ron was updated to include a new `ConnTest` command struct that
encapsulates the endpoint to test against, how long to wait, and what
UDP packet to send (if necessary).

miniccc was updated to include a handler for the new `ConnTest` command
that simply tries to dial the endpoint (in the case of TCP), and if
necessary write the UDP packet to the socket (in the case of UDP).

minimega was updated to include support for a new "cc test-conn"
command, as well as adding a "connectivity" column to the "cc commands"
table.

Documentation for minimega's "cc" command was also updated to include
details and examples of how to use the new "test-conn" command.