[minimega] Prevent dnsmasq from leaking DNS results (#1471)
* [minimega] Prevent dnsmasq from leaking DNS results

This commit updates the command line flags for dnsmasq to explicitly
point to a non-existent resolv.conf file rather than letting it use the
default /etc/resolv.conf. This prevents leakage of DNS results as
described in #1421.

This commit also adds an option to the `dns configure` command for
adding upstream DNS servers. Adding an upstream server simply adds a
nameserver entry to the (initially non-existent) resolve.conf file
dnsmasq is configured to use. Upstream servers added will immediately be
recognized by dnsmasq since it polls the file for changes.

closes #1421

* Add help docs for dnsmasq upstream server options