[minimega] Add "vm net bond" API command
Bond two or more network interfaces together for a VM. All interfaces
being bonded must be on the same bridge or the command will fail. If all
the interfaces being bonded are not on the same VLAN, the bond uses the
VLAN from the first interface. If at least one of the bonded interfaces
is configured for QinQ, the bond will be configured for QinQ as well.