minimega: add CLI flag for setting default VLAN range for namespaces
This is useful for avoiding VLAN ID ranges that are used for production
or lab networks that minimega may be tied into for hardware-in-the-loop
access to real devices/workstations/servers.

The default for the flag is the same as the current default range, so
not setting the flag provides backwards compatibility.