[minimega] add -recovery flag to recover VMs after minimega crash
NOTE: this only works for KVM VMs for now (not containers)

If the `-recovery` flag is provided, the `-force` flag is not (the
`-force` flag will take precedence), and the minimega unix socket still
exists (e.g., after minimega crashing for some reason), minimega will
attempt to recover any QEMU KVM VMs still running that are mapped to
valid minimega namespaces and VMs per the myriad of files that minimega
writes to `/tmp/minimega`.

When minimega creates a QEMU KVM VM, it sets the `-name` option to
minimega's VM ID and sets the `-uuid` option to the UUID generated or
specified for the minimega VM. These two `qemu-system` process flags,
along with the myriad of files present in `/tmp/minimega` are used to
piece together which VMs minimega started, in what namespaces, with what
taps an VLANs, etc. minimega then uses this information to (re)populate
its internal data structures such that commands like `vm info` show the
correct data. It also uses this information to reconnect VM VNC and CC
connections.