ron: add option for only executing a command once (#1508)
This allows, for example, the `shutdown -r now` command to be sent to a
VM using command and control without putting the VM into a perpetual
reboot loop. This is a contrived example, since a minimega VM can be
rebooted without using command and control, but it still gets the point
across. There may be other commands that are sent via command and
control that end up causing a VM to reboot that would benefit from this
as well.

New cc commands added:

  cc exec-once command...
  cc background-once command...